# Optimize Keyword.new/1,2 functions