Follow-up on https://github.com/elixir-lang/elixir/pull/3176#discussion_r26907713.
I think we do not need hand-written implementation, the `put_new` is well suited.
The timings for 200 elements enum:
`Keyword.new/1`:
```
new:      50000   60.69 µs/op
old:      10000   124.75 µs/op
```
`Keyword.new/2`:
```
new:      50000   55.86 µs/op
old:       1000   1113.29 µs/op
```